Refactor protocol imports and enhance _Node definition:
Updated import paths for _Node across multiple files to streamline the
codebase. Expanded the _Node protocol in protocol.py to define its role
as a minimal graph vertex, ensuring compatibility with pipeline
execution. Removed the deprecated _Node definition from operator
protocol.
