+++ miv/core/operator/protocol.py
@@ -3,13 +3,11 @@ Specification of the behaviors for Operator modules.
 """
 
 from typing import Protocol, Any
-
-import pathlib
 from abc import abstractmethod
 
 from .policy import RunnerBase
-from ..protocol import _Loggable, _Chainable, _Cachable
-from ..datatype import DataTypes
+
+__all__ = ["_Runnable"]
 
 
 class _Runnable(Protocol):
@@ -24,25 +22,3 @@ class _Runnable(Protocol):
 
     @abstractmethod
     def __call__(self, *args: Any, **kwargs: Any) -> Any: ...
-
-
-class _Node(
-    _Loggable,
-    _Runnable,
-    _Chainable,
-    _Cachable,
-    Protocol,
-):
-    """
-    Protocol defining the complete behavior of an Operator node.
-    Each protocol aspect is separately defined to allow for cleaner composition.
-    """
-
-    analysis_path: str
-
-    def output(self) -> DataTypes: ...
-    def set_save_path(
-        self,
-        path: str | pathlib.Path,
-        cache_path: str | pathlib.Path | None = None,
-    ) -> None: ...
